homework-7
first step
